requirements.txt
# ================================
# CORE BACKEND
# ================================
fastapi==0.110.0
uvicorn[standard]==0.27.1

# ================================
# FILE UPLOAD & FORM HANDLING
# ================================
python-multipart==0.0.9

# ================================
# DATA & UTILITIES
# ================================
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
orjson==3.9.15

# ================================
# AI / MACHINE LEARNING
# ================================
numpy==1.26.4
scipy==1.12.0
torch>=2.0.0
torchaudio>=2.0.0

# Whisper (Speech to Text)
openai-whisper==20231117

# ================================
# AUDIO / VIDEO PROCESSING
# ================================
ffmpeg-python==0.2.0
pydub==0.25.1

# ================================
# TEXT / NLP SUPPORT
# ================================
regex==2023.12.25

# ================================
# HTTP CLIENT (API CALLS TO AI)
# ================================
requests==2.31.0
httpx==0.27.0

# ================================
# ASYNC / PERFORMANCE
# ================================
anyio==4.3.0
async-timeout==4.0.3

# ================================
# LOGGING & DEBUG
# ================================
loguru==0.7.2

# ================================
# OPTIONAL (FUTURE READY)
# ================================
websockets==12.0
jinja2==3.1.3
